♻️ refactor(ui): collect domain-agnostic blocks under ui/
Drawer, Dropdown, Grid, SortableList, Tooltip and icons never referenced the
AudioPlayer domain, so they belonged to the same layer as StyledBtn and
CssTransition. Splitting them across components/ and ui/ left the placement
rule as "how big is it", which had already broken down — Tooltip.tsx (30
lines) sat in components/ while StyledBtn.tsx (18 lines) sat in ui/.

Move them into ui/ so the rule becomes "does it know the domain", answered
by the folder alone: ui/ imports no domain code, components/ holds only
AudioPlayer.

Controller/Icon.tsx moves along as ui/IconSlot.tsx. It is a slot that picks
customIcon over a default, not an icon, and the old name sat one letter away
from the icons/ set. Being a sibling of icons/ now, it keeps its
ReactElement<SvgIconProps> typing instead of loosening to ReactNode.

Behavior unchanged: 451 tests pass, typecheck and lint clean.
