+++ package/src/components/AudioPlayer/Interface/Controller/PlayList/index.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
-import Drawer from "@/components/Drawer";
-import Grid, { GridItemLayoutProps } from "@/components/Grid";
+import Drawer from "@/ui/Drawer";
+import Grid, { GridItemLayoutProps } from "@/ui/Grid";
 import { useUIContext } from "@/components/AudioPlayer/Context/hooks/useUIContext";
 import { useResolvedGridArea } from "../../hooks/useResolvedGridArea";
 import { PlayList } from "./Content";
